# openapi/openapi.yaml
openapi: 3.1.1
info:
  title: MarketStat API
  version: 1.0.0
  description: API for analyzing salaries in the Russian labor market.
servers:
  - url: http://localhost:5000/api/v1
    description: MarketStat API v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequestDto:
      type: object
      properties:
        username:
          type: string
          example: "analyst_user"
        password:
          type: string
          format: password
          example: "password123"
      required: [username, password]
    RegisterUserDto:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
          example: "new_analyst"
        password:
          type: string
          minLength: 8
          maxLength: 100
          format: password
          example: "password123"
        email:
          type: string
          format: email
          maxLength: 255
          example: "new_analyst@company.com"
        fullName:
          type: string
          maxLength: 255
          example: "Analyst"
        isAdmin:
          type: boolean
          default: false
      required: [username, password, email, fullName]
    UserDto:
      type: object
      properties:
        userId:
          type: integer
          format: int32
          example: 12
        username:
          type: string
          example: "analytst_user"
        email:
          type: string
          format: email
          example: "analyst@company.com"
        fullName:
          type: string
          example: "Analyst"
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
      required: [userId, username, email, fullName, isActive, createdAt]
    AuthResponseDto:
      type: object
      properties:
        token:
          type: string
          example: "token_123"
        expiration:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserDto'
      required: [token, expiration, user]
    PartialUpdateUserDto:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
          minLength: 1
          maxLength: 255
          example: "Jane R. Doe"
        email:
          type: ["string", "null"]
          format: email
          maxLength: 255
          example: "jane.doe@newemail.com"

    # DimDate
...
    ErrorPayload:
      type: object
      properties:
        message:
          type: string
          description: A human-readable message describing the error.
          example: "Invalid DateId."
      required:
        - message
  responses:
    BadRequestResponse:
      description: Bad Request - The request is malformed, contains invalid syntax, or fails validation (maps to ValidationException or ArgumentException).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
    UnauthorizedResponse:
      description: Unauthorized - The request requires authentication, and the JWT is missing or invalid (maps to AuthenticationException).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
...
    NotFoundResponse:
      description: Not Found - The requested resource does not exist (maps to NotFoundException).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
    ConflictResponse:
      description: Conflict - The request could not be completed due to a conflict with the current state of the resource (maps to ConflictException).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      description: Creates a new user account. Public endpoint.
      security: []
      requestBody:
        description: The registration details for the new user.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserDto'
      responses:
        '201':
          description: User registered successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
        '500':
          description: Internal server error.
  /auth/login:
    post:
      tags:
        - Auth
      summary: Logs in an existing user
      description: Authenticates a user and returns a JWT token and user details. This endpoint is public.
      security: []
      requestBody:
        description: The login credentials for the user.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
      responses:
        '200':
          description: Authentication successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          description: Internal server error.
  
  # /auth/profile/me:
  #   patch:
  #     tags:
  #       - Auth
  #     summary: Partially updates the current user's profile
  #     description: "Allows an authenticated user to update their own profile information (FullName, Email). The user is identified by their JWT."
  #     security:
  #       - bearerAuth: []
  #     requestBody:
  #       description: The fields to update on the user's profile.
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: '#/components/schemas/PartialUpdateUserDto'
  #     responses:
  #       '200':
  #         description: The user profile was updated successfully.
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/UserDto'
  #       '400':
  #         $ref: '#/components/responses/BadRequestResponse'
  #       '401':
  #         $ref: '#/components/responses/UnauthorizedResponse'
  #       '404':
  #         $ref: '#/components/responses/NotFoundResponse'
  #       '409':
  #         description: "Conflict - The request could not be completed (e.g., the new email is already taken)."
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/ErrorPayload'
  
  # DimDate
...